% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrm_basin_stats.R
\name{gen_upstream_basin}
\alias{gen_upstream_basin}
\title{Provided a digital elevation model (DEM), X-Y coordinates (in map units), using RSAGA function uses flow tracing 
algorithms to determine the upstream contributing area. Function returns path to binary grid defining
the upstream catchment for the provided coordinates.}
\usage{
gen_upstream_basin(
  dem,
  x_coord,
  y_coord,
  method = 2,
  converge = 1.1,
  outdir,
  basin_id,
  env
)
}
\arguments{
\item{dem}{(character) Filepath to digital elevation model, assumed to be hydrologic conditioned.}

\item{x_coord}{(flaot) The X coordinate in map units, must be withing extent of DEM.}

\item{y_coord}{(flaot) The Y coordinate in map units, must be withing extent of DEM.}

\item{method}{(integer) [0] Deterministic 8; [1] Deterministic Infinity;[2] Multiple Flow Direction (default)}

\item{converge}{(flaot) Convergence factor for Multiple Flow Direction algorithm (default 1.1)}

\item{outdir}{(character) Filepath for binary grid output.}

\item{basin_id}{(integer) A unique basin identifier, will be appended to output binary grid name.}

\item{env}{A RSAGA environment,i.e., RSAGA::rsaga.env()}
}
\value{
(character) Filepath to output binary upstream catchment grid, will match extent of DEM.
}
\description{
Provided a digital elevation model (DEM), X-Y coordinates (in map units), using RSAGA function uses flow tracing 
algorithms to determine the upstream contributing area. Function returns path to binary grid defining
the upstream catchment for the provided coordinates.
}
