% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrm_basin_stats.R
\name{get_basin_stats}
\alias{get_basin_stats}
\title{Provided a zonal raster and input layers, using RSAGA this function will return a table
of zonal statistics as a data.frame.}
\usage{
get_basin_stats(zone_rast, catlist = NULL, statlist, outdir, basin_id, env)
}
\arguments{
\item{zone_rast}{(character) Filepath to SAGA grind defining the zones to analyse, acts as a mask,i.e., output from 'gen_upstream_basin'.}

\item{catlist}{(character; optional) Filepath(s) (Seperated by ';') used to delineate unique condition units (UCUs). (default NULL).}

\item{statlist}{(character) Filepath(s) (Seperated by ';') to SAGA grids with continuous data, statistics are calculated for each grid.}

\item{outdir}{(character) Filepath to location for SAGA statistics table to be written.}

\item{basin_id}{(integer) A unique basin identifier, will be appended to output statistics table name.}

\item{env}{A RSAGA environment,i.e., RSAGA::rsaga.env()}
}
\value{
A data.frame of statistics for each input grid, by zone.
}
\description{
Provided a zonal raster and input layers, using RSAGA this function will return a table
of zonal statistics as a data.frame.
}
