% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrm_basin_stats.R
\name{hydro_condition_dem}
\alias{hydro_condition_dem}
\title{Provided a DEM function will apply hydrologic pre-processing outputting a 
hydrologic conditioned DEM, free of sinks, using methods defined in Wang, L. & H. Liu (2006).}
\usage{
hydro_condition_dem(dem, outdir, env, minslope = 0.1)
}
\arguments{
\item{dem}{(character) Filepath to input digital elevation model (DEM).}

\item{outdir}{(character) Filepath for output hydrologic conditioned DEM, will be named 'FILLED.sgrd'}

\item{env}{A RSAGA environment,i.e., RSAGA::rsaga.env()}

\item{minslope}{(float) Minimum slope gradient (degrees) to preserve from cell to cell; with a value of zero
sinks are filled up to the spill elevation (which results in flat areas).}
}
\value{
(character) File path to FILLED.sgrd.
}
\description{
Provided a DEM function will apply hydrologic pre-processing outputting a 
hydrologic conditioned DEM, free of sinks, using methods defined in Wang, L. & H. Liu (2006).
}
