% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_clean_polygon.R
\name{raster_to_clean_polygon}
\alias{raster_to_clean_polygon}
\title{Polygonize a classified raster layer, dropping crumbs and filling holes.}
\usage{
raster_to_clean_polygon(
  r,
  min_area_drop_m2,
  max_area_fill_m2,
  smooth,
  smooth_method
)
}
\arguments{
\item{r}{Classified Raster object, STARS object or Spatial Grid.}

\item{min_area_drop_m2}{Numeric. Minimum area in square meters for which polygons will be dropped}

\item{max_area_fill_m2}{Numeric. Maximum area in square meters of internal polygon holes for which
holes will be remove.}

\item{smooth}{Boolean. Should the resulting polygons geometry be simplified.}

\item{smooth_method}{String. Smoothing method to pass to \code{\link{smoothr}},
smooth must be TRUE.}
}
\value{
Polygonized sf object representing the input raster categories.
}
\description{
Must provide a classified raster, minimum area for which polygons should
be dropped (meteres sqaured), and maximum area for which polygon holes
should be filled. Optionally, polygons may be simplified using \code{\link{smoothr}} 
function, a smoothing method must be specified.
}
\examples{
\dontrun{
#Create a classified *stars* object 
r <- stars::read_stars('CategoryPredictionMap.tif')

#Polygonize classified raster
p <- raster_to_clean_polygon(r,2500,2500,TRUE,'chaikin')
}
}
