% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrm_basin_stats.R
\name{run_basin_stats}
\alias{run_basin_stats}
\title{Provided 'set_grass_env' has been called, function returns upstream
basin statistics for all pour points defined in a pour point data
frame.}
\usage{
run_basin_stats(
  procs = 1,
  proc_type = "FORK",
  pour_pnts,
  covar_rast,
  stat_vect
)
}
\arguments{
\item{procs}{Number of concurrent process to use, defaults to 1.}

\item{proc_type}{Type of 'parallel' cluster to apply, defaults to 'FORK'}

\item{pour_pnts}{A data.frame with three columns labeled 'X', 'Y' and 'UID', 
representing the X, and Y coordinates of the pour point in units matching 
the DEM, and a unique identifying number (UID).}

\item{...}{
  Arguments passed on to \code{\link[=upstream_basin_stats]{upstream_basin_stats}}
  \describe{
    \item{\code{covar_rast}}{Character vector of GRASS-GIS covariate
raster names (assumed to be present in GRASS env.) 
over which to calculate upstream statistics.}
    \item{\code{stat_vec}}{Character vector, order corresponds 
to 'covar_rast' and specifies what upstream statistic
to return for each covariate raster. Must be one of 
'N', 'MIN', 'MAX', 'RANGE', 'MEAN', 'MAE', 'STDDEV', 
'VAR', 'SUM', 'PRCT' or 'NULL_CELLS'.}
  }}
}
\value{
A data frame with upstream statistics computed for each covarite raster
defined in 'covar_rast' over each pour point provided in 'pour_pnts', along with
the pour point UID.
}
\description{
This function assumes 'set_grass_env' has been called. This function 
iterates 'upstream_basin_stats' over a table of pour points
in parallel. The table of pour points must be a data.frame with
three columns labeled 'X','Y' and 'UID', representing the X, and
Y coordinates of the pour point in units matching the DEM, and a 
unique identifying number (UID). A this function runs in parallel,
the user must specify the number and type of processes to apply.
}
