% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{stack_rasters}
\alias{stack_rasters}
\title{Create a stack of rasters}
\usage{
stack_rasters(rasters, target_raster, outdir = NULL)
}
\arguments{
\item{rasters}{filename(s) (character) of raster files to be
stacked with the target_raster. Input rasters can be any format read by
\code{raster}}

\item{target_raster}{filename (character) of target raster used to align all
other rasters.}

\item{outdir}{optional output folder path (character) where .img files of the stacked
rasters are saved.}
}
\value{
RasterStack object
}
\description{
Create a stack of rasters
}
