% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrm_basin_stats.R
\name{upstream_basin_stats}
\alias{upstream_basin_stats}
\title{Provided 'set_grass_env' has been called, function returns upstream
basin statistics for a user specified pour point, called from 'run_basin_stats'.}
\usage{
upstream_basin_stats(x, y, uid, covar_rast, stat_vec)
}
\arguments{
\item{x}{The X-Coordinate of the pour point in units of DEM.}

\item{y}{The Y-Coordinate of the pour point in units of DEM.}

\item{uid}{A unique identifier number for the pour point.}

\item{covar_rast}{Character vector of GRASS-GIS covariate
raster names (assumed to be present in GRASS env.) 
over which to calculate upstream statistics.}

\item{stat_vec}{Character vector, order corresponds 
to 'covar_rast' and specifies what upstream statistic
to return for each covariate raster. Must be one of 
'N', 'MIN', 'MAX', 'RANGE', 'MEAN', 'MAE', 'STDDEV', 
'VAR', 'SUM', 'PRCT' or 'NULL_CELLS'.}
}
\value{
A vector with upstream basin statistics for 
each covarite raster defined by 'covar_rast' and 'stat_vec'.
}
\description{
This function assumes 'set_grass_env' has been called. This function will
for a specified pour point delineate the upstream area from the 
DEM present in the GRASS environment using 'r.water.outlet'. Then for
all of the GRASS covariate raster defined by the 'covar_rast' basin statistics
are calculated using 'r.univar'. For each covariate raster the statistic to
return must be specified in the corresponding 'stat_vec' character vector. Returns a 
vector with the UID and associated statistics for each covariate raster.
Function called by 'run_basin_stats' so need not be called by user.
}
