% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmap_processing.R
\name{wetland_model}
\alias{wetland_model}
\title{Run model and diagnostics for wetland prediction}
\usage{
wetland_model(qdatafn, model.type, model.folder, unique.rowname, predList,
  predFactor, response.name, response.type, seed, MODELfn = NULL,
  na.action = "na.roughfix", ...)
}
\arguments{
\item{qdatafn}{training data filename.}

\item{model.type}{model type to use (string). One of "RF", "QRF", or "CF".}

\item{model.folder}{folder where the output folder will be created.}

\item{unique.rowname}{unique identifier column name in the training data.}

\item{predList}{character vector of predictor names that match the column
names in the training data.}

\item{predFactor}{character vector of predictor names that are factors (categorical), or
FALSE if there are no categorical predictors.}

\item{response.name}{response variable (string); column name in the training data.}

\item{response.type}{response type (string). One of "binary", "categorical", or "continuous".}

\item{seed}{number (integer) used to initialize randomization.}

\item{MODELfn}{optional filename to use for saved model output. If \code{MODELfn = NULL} (default)
a unique name is created automatically using the date and time the model is run.}

\item{na.action}{optional string to specify action to take if there are NA values
in the predictor data. Defaults to \code{na.roughfix}.}

\item{...}{any other parameters to pass to ModelMap::model.build.}
}
\value{
list containing model object from ModelMap::model.build and output model folder string.
}
\description{
Run model and diagnostics for wetland prediction
}
