% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlandmapR.R
\docType{package}
\name{wetlandmapR}
\alias{wetlandmapR}
\title{wetlandmapR: Mapping Wetland Ecosystems}
\description{
Tools for mapping wetland ecosystems using data driven R statistical methods
like Random Forests and open source GIS.
}
\details{
This package provides tools for running the \code{\link[ModelMap]{model.build}}
and \code{\link[ModelMap]{model.mapmake}} functions, specifically for
modelling and mapping wetland ecosystems. Additional functions help generate
the necessary input training data and raster look up table inputs.

Wetland models can be run using area of interest polygons, restricting output
to specific drainage basins, for example.

wetlandmapR depends on RSAGA for some raster processing. RSAGA depends on
\href{http://www.saga-gis.org/en/index.html}{SAGA GIS} being installed and
accessible on your computer. Please see the RSAGA documentation for
instructions on how to do this.
}
