% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{create_dem_products}
\alias{create_dem_products}
\title{Create products from a DEM using RSAGA}
\usage{
create_dem_products(
  dem,
  stream_vec = NULL,
  burn_val = NULL,
  outdir,
  products = NULL,
  param_list = NULL,
  param_values = NULL
)
}
\arguments{
\item{dem}{filename (character) of input DEM raster.}

\item{stream_vec}{(Optional) filename (character) to existing vector stream network to burn into DEM, e.g., Fresh Water Atlas.}

\item{burn_val}{(Optional) double. The value of Epsilon for burning stream network into DEM.}

\item{outdir}{output folder path (character).}

\item{products}{character. The DEM products to be created. Can be a vector
with any of the following: c("SLOPE", "ASPECT", "DAH", "MRVBF", "TPI",
"CPLAN", "CPROF", "TOPOWET", "CAREA"). If not provided, this parameter
defaults to generating all products.}

\item{param_values}{This is a list of lists corresponding to param_vect. For each
parameter specified in 'param_vect', the additional parameters for the corresponding
SAGA module must be specified using a list of those additional parameters and their values.}

\item{param_vect}{character. Similar to the products parameter, this is a vector of
the SAGA module names (specified in products parameter) for which non-default
parameters will be passed. Defaults to NULL.}
}
\description{
Creates raster derivatives (products) from an input Digital Elevation Model
(DEM) using SAGA-GIS.
}
\details{
This function uses RSAGA which requires having SAGA-GIS (v2.3+) installed.
The input DEM can be in any format that can be read by \code{raster::raster}.
Output rasters are saved in the SAGA Grid format, including a SAGA Grid
version of the input DEM, saved as "ELEV.sgrd".

The following products can be created by using the \code{products} parameter
(if \code{products = NULL} all products will be created by default):
  SLOPE:   Slope (percent).
  ASPECT:  Aspect (degrees).
  DAH:     Diurnal Anisotrophic Heating.
  MRVBF:   Multiresolution Index of Valley Bottom Flatness.
  TPI:     Topographic Position Index.
  CPLAN:   Plan Curvature.
  CPROF:   Profile Curvature.
  TOPOWET: SAGA Wetness Index.
  CAREA:   Top-Down Flow Accumulation.
}
\examples{
\dontrun{
create_dem_products(dem = "data/dem.tif", outdir = "output",
                    products = c("SLOPE", "CAREA"))
}
}
