% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{create_dem_products}
\alias{create_dem_products}
\title{Create products from a DEM using RSAGA}
\usage{
create_dem_products(dem, outdir, products = "ALL")
}
\arguments{
\item{dem}{filename (character) of input DEM raster.}

\item{outdir}{output folder path (character).}

\item{products}{character. The DEM products to be created. Can be a vector
with any of the following: c("SLOPE", "ASPECT", "DAH", "MRVBF", "TPI", "CPLAN",
"CPROF", "TOPOWET", "CAREA")}
}
\description{
Create products from a DEM using RSAGA
}
