% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{grid_values_at_sp}
\alias{grid_values_at_sp}
\title{Extract raster values at points}
\usage{
grid_values_at_sp(x, y, filename = NULL)
}
\arguments{
\item{x}{Raster* object}

\item{y}{SpatialPoints* object}

\item{filename}{optional output CSV filename}
}
\value{
SpatialPoints* object with values added to it's data frame
}
\description{
Extract raster values at points
}
