% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{stack_rasters}
\alias{stack_rasters}
\title{Create a stack of rasters}
\usage{
stack_rasters(
  rasters,
  aligned = FALSE,
  target_raster = NULL,
  outdir = NULL,
  rastLUTfn = NULL
)
}
\arguments{
\item{rasters}{filename(s) (character) of raster files to be
stacked with the target_raster. Input rasters can be any format read by
\code{raster::raster}.}

\item{aligned}{optional boolean, indicates whether input rasters are already
aligned to the same grid. Use \code{aligned = TRUE} if \code{stack_rasters}
has already been run and you want to create a RasterStack object from an
existing list of aligned rasters. If \code{aligned = FALSE} (default),
\code{target_raster} must be provided.}

\item{target_raster}{optional filename (character) of target raster used to
align all other rasters. Required if \code{aligned = FALSE}.}

\item{outdir}{optional output folder path (character) where .img files of the
stacked rasters are saved. If no value is provided, no .img files are saved.
If \code{aligned = TRUE}, no output .img files are saved, even if an output
folder is given.}

\item{rastLUTfn}{optional filename (character) of output rastLUT .csv file,
for use in \code{wetland_map}.}
}
\value{
RasterStack object
}
\description{
Aligns input raster(s) to a target raster so that extent, cell size, and
cell origin are the same, returning a RasterStack object.
}
\details{
This function will reproject the input rasters (using bilinear resampling)
to the same projection as the target if necessary.

If \code{outdir} is given an ERDAS Imagine .img file is output for each
aligned raster.

This function can also output a "rastLUT" CSV file for use as input to
\code{\link{wetland_model}}, where the raster names are used as the predictor
name.

If you have a set of rasters that are already aligned, you can use this
function to create a RasterStack without creating duplicate output rasters
by using the \code{aligned} parameter.
}
\examples{
\dontrun{
# Create raster list from output files generated by create_dem_products()
raster_list <- list.files("output", "sdat$", full.names = TRUE)

raster_stack <- stack_rasters(rasters = raster_list,
                              target_raster = "data/dem.tif",
                              outdir = "output",
                              rastLUTfn = "output/rastLUT.csv")

# Stack existing, aligned, rasters
raster_list <- c("output/ELEV.img", "output/SLOPE.img", "output/ASPECT.img")
raster_stack_2 <- stack_rasters(rasters = raster_list,
                                aligned = TRUE)
}
}
