% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_stack.R
\name{stack_rasters}
\alias{stack_rasters}
\title{Create a stack of input rasters}
\usage{
stack_rasters(
  rasters,
  target_raster,
  outdir = NULL,
  rastLUTfn = NULL,
  rastNames = NULL,
  NAval = NaN,
  bands = 1,
  aligned = FALSE
)
}
\arguments{
\item{rasters}{filename(s) (character) of raster files to be
stacked with the target_raster. Input rasters can be any format read by
\code{terra::rast}.}

\item{target_raster}{filename (character) of target raster used to
align all other rasters.}

\item{outdir}{(optional) output folder path (character) where .tif files of the
stacked rasters are saved. If no value is provided, no .tif files are saved. Outputs will be
placed in a directory names 'resampl'.}

\item{rastLUTfn}{(optional) Directory path (character) for output 'rastLUT.csv' file,
for use in \code{wetland_map}. Default NULL.}

\item{rastNames}{(optional) (character) If rastLUTfn is provided, must provide 
corresponding names of each raster in 'rasters'.}

\item{NAval}{NA value of input raster if not NaN (must be same for all), defaults to NaN.}

\item{bands}{(optional) character vector corresponding to rastLUT indicating which band to use as model input.
Defaults to first band for each layer specified in rastLUT.}

\item{aligned}{(Boolean) If true, function acts as wrapper for terra::rast, default false.}
}
\value{
Multiband Terra raster object
}
\description{
Aligns input raster(s) to a target raster so that extent, cell size, and
cell origin are the same, returning a Terra raster object.
}
\details{
This function will reproject and or resample the input rasters (using bilinear resampling)
to the same projection as the target, if necessary.

If \code{outdir} is given an GTIFF (.tif) file is output for each
aligned raster.

This function can also output a "rastLUT" CSV file for use as input to
\code{\link{wetland_model}}, where the raster names are used as the predictor
name.

If you have a set of rasters that are already aligned, you can use this
function to create a RasterStack without creating duplicate output rasters
by using the \code{aligned} parameter.
}
\examples{
\dontrun{
# Create raster list from output files generated by create_dem_products()
raster_list <- list.files("output", "sdat$", full.names = TRUE)

raster_stack <- stack_rasters(rasters = raster_list,
                              target_raster = "data/dem.tif",
                              outdir = "output",
                              rastLUTfn = "output/rastLUT.csv")

}
}
