% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmap_processing.R
\name{wetland_map}
\alias{wetland_map}
\title{Wetland map production}
\usage{
wetland_map(model.out, model.folder, rastLUTfn, aoi = NULL,
  aoi.col = NULL, na.action = "na.omit", ...)
}
\arguments{
\item{model.out}{list object returned from \code{wetland_model}. List
contains model object(s) from \code{\link[ModelMap]{model.build}} for each
AOI; list elements are named by the output model folder string (including
AOI target name if aoi.target is provided).}

\item{model.folder}{folder where the output from \code{wetland_model} was
created. Same as \code{model.folder} input to \code{wetland_model}.}

\item{rastLUTfn}{filename of a .csv or dataframe of a rastLUT.}

\item{aoi}{optional SpatialPolygon object that was used to intersect with
input points. This parameter is required if \code{model.out} was generated
using an AOI.}

\item{aoi.col}{optional string to specify the column name in the \code{aoi}
data which identifies seperate AOIs. Required if \code{aoi} is provided.}

\item{na.action}{optional string to specify action to take if there are NA
values in the prediction data. Defaults to \code{na.omit}.}

\item{...}{any other parameters to pass to \code{ModelMap::model.mapmake}.}
}
\description{
This function runs \code{\link[ModelMap]{model.mapmake}} to generate raster
prediction surfaces using model output from \code{\link{wetland_model}}.
}
\details{
The name of the model object, output from \code{\link{wetland_model}} is used
to determine the output folder where the ERDAS Imagine .img output raster is
saved.

If an AOI was used to generate the model object(s), the same AOI
SpatialPolygon object can be provided as input to this function so that
the output raster is produced for the corresponding AOI. If a single model
object has been generated for a specific AOI, that model can be applied to a
larger extent by not providing an AOI input to this function. If an AOI is
provided, the rasters listed in \code{rastLUT} are clipped to the AOI
polygon used to generate the model.

Temporary rasters and temporary rastLUTs specific to each AOI are saved in
the output folder; these temporary files are deleted once the prediction
surface has been created.
}
\examples{
\dontrun{
# RastLUT:
rastLUT <- read.csv("data/rastLUT.csv",
                    header = FALSE,
                    stringsAsFactors = FALSE)

# Character vector of the predictor names, used as input to the model:
predList <- rastLUT[, 2]

# Training data, that has been intersected with an AOI layer:
qdatafn <- "data/training_points_attd.csv"

# Run the model
model.out <- wetland_model(qdatafn = qdatafn,
                           model.type = "RF",
                           model.folder = "./output",
                           unique.rowname = "OBJECTID",
                           predList = predList,
                           predFactor = FALSE,
                           response.name = "T_W_Class",
                           response.type = "categorical",
                           seed = 44,
                           response.target = c("Wb", "Wf", "Ws"),
                           aoi.col = "BASIN")

# AOI layer:
aoi_polys <- raster::shapefile("data/aoi.shp")

wetland_map(model.out = model.out,
            model.folder = "./output",
            rastLUTfn = rastLUT,
            aoi = aoi_polys,
            aoi.col = "BASIN")
}
}
