% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmap_processing.R
\name{wetland_map}
\alias{wetland_map}
\title{Wetland map production}
\usage{
wetland_map(model.obj, folder, MODELfn, rastLUTfn, na.action = "na.omit",
  ...)
}
\arguments{
\item{model.obj}{model object returned from \code{wetland_model}.}

\item{folder}{folder name used for output}

\item{MODELfn}{filename to use for saved model output.}

\item{rastLUTfn}{filename of a .csv for a rastLUT.}

\item{na.action}{optional string to specify action to take if there are NA
values in the prediction data. Defaults to \code{na.omit}.}

\item{...}{any other parameters to pass to ModelMap::model.mapmake.}
}
\description{
Wetland map production
}
